HEMCO Update: Add Git repository status to the Configure header
In the top-level HEMCO/CMakeLists.txt, we add CMakeScripts to the search
path before printing the configure header.  This is so that we can
call function gc_repo_version to print the Git status out to
the configure message.

In this commit, the GCHPctm wrapper now includes this commit
from the src/HEMCO submodule:

3b0a804 Add Git repository status to the Configure header

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>