GCHP rundir creation bug fix; Move main.F90 to Interfaces/GCClassic
(1) We needed to add a new case to for $sim_extra_option=none in the
script run/shared/setupConfigFiles.sh, so that the "@" characters in
the HISTORY.rc.fullchem template file will be removed.

(2) We have also moved the GEOS-Chem "Classic" driver program main.F90
from GeosCore to Interfaces/GCClassic.  This usage is now more consistent
with how we interface the GEOS-Chem science codebase to other
external modeling contexts such as NASA/GEOS GCM and GCHPctm.

(3) We also now tell the src/CMakeLists.txt to look for main.F90 directly
 with the path GEOS-Chem/Interfaces/GCClassic/main.F90, instead of
relying upon a symbolic link.

(4) Also updated the CMakeLists in the GCClassic, GEOS-Chem, and HEMCO
folders to display the repository status during the configure state.
This is done using the get_repo_version CMake function, which is
included in GC-Helpers.cmake and HCO-Helpers.cmake.

With this commit, the GCHPctm wrapper has been informed about these
commits to the src/GEOS-Chem submodule:

097266a7 GCHP bug fix: Also remove "@" characters for standard simulations
f603b110 Move main.F90 to Interfaces/GCClassic; Display Git status in configure

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>