Change default run directory path to two levels up from install
This update restores previous behavior for using the following
commands to build GCHP from a run directory:

   mkdir -p build; cd build; cmake ../CodeDir; make -j; make install

This allows the recommended procedure for building GCHP to match that of
GEOS-Chem Classic. Changes to the recommended procedure in the future
will be applied to both at the same time so that a common set of
instructions may always be used.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>