GEOS-Chem submod update: Add data entries for TOMAS seasalt & dust to HEMCO config templates
This commit informs the GCClassic superproject about the following
commits that were made to the src/GEOS-Chem submodule:

57e45091 Add data entries for TOMAS size-resolved seasalt & dust extensions

We had to copy the DustDead and Seasalt extension data entries in
HEMCO_Config.rc for the TOMAS extensions, assigning them the proper
extension numbers.  This fixed an issue in which the TOMAS simulations
could not find certain data files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>