Disable CI-pipeline quick run test and bring back quick build
GCHP continuous integration quick tests currently require run directory
updates in submodule MinimumStandaloneTransportTracer. This is causing
problems because its run directory config files are getting out of date
as GEOS-Chem development progresses. Updating the files in two separate
repositories is cumbersome and easy to forget. We will therefore bypass
the quick test until it can be modified to tap into GEOS-Chem repo files.

The GCHP quick build test is now added back in to check that each bugfix
and dev branch commit compiles. This replaces the quick test temporarily.
It matches the quick build test previously used for GCHP during 13.0.0
development.

Also in this commit: remove setting LD_LIBRARY_PATH in quick-test.yml.
This is no longer needed in GCHP and is removed in preparation for when
we turn the quick test back on.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>