Apply MAPL kludge to fix incorrect lev positive attribute in History files
History collection output files now have lev positive 'up' rather than
'down' for all collections except emissions. Please note that this is
a quick and dirty fix that will be replaced with a more elegant and
robust solution in the future. Also note that adding collections for
non-GEOS-Chem gridcomps or HEMCO emissions will require further edits to
MAPL to ensure the positive attribute is correct. For these cases
positive should be 'down' unless the data arrays are vertically flipped.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>