Further fixes for compiling with -DLUO_WETDEP=on
The previous fix for LUO_WETDEP still resulted in the following in
CMakeCache.txt:

   //No help, variable specified on the command line.
   LUO_WETDEP:UNINITIALIZED=on

To resolve this, we now follow what is done for the RRTMG option. That is,
set LUO_WETDEP at the top of src/GCHP_GridComp/GEOSChem_GridComp/CMakeLists.txt.
The line for LUO_WETDEP passed to target_compile_definitions has been
moved to the section for the other GEOS-Chem build properties instead of
the section for compiler options. This fix now results in the following
output in CMakeCache.txt when building with -DLUO_WETDEP=on:

   //Switch to build the Luo et al (2020) wetdep scheme into GEOS-Chem
   LUO_WETDEP:BOOL=on

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>