GEOS-Chem submod update: Increase drydep resistance for O3 on ice/snow
This commit informs the GCHP superproject about the following
commits that were pushed to the Github geoschem/geos-chem repository:

49d10fa17 Merge branch 'feature/O3-ice-drydep' of ssh://github.com/r-pound/geos-chem into feature/O3-ice-drydep

This update is tagged with 13.4.0-alpha.8.  The alpha.6 and alpha.7
benchmarks (for sulfur chemistry via KPP) revealed large changes that
need to be further quantified.  In order to not slow down the
development process, we have leapfrogged to alpha.8.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>