GEOS-Chem submod update: Merge branch 'feature/Hg-with-kpp' into dev
This commit brings the Hg simulation implemented as a KPP chemistry
mechanism on top of the 13.4.0-alpha.11 tag.

NOTE: The -DCUSTOMMECH=y option has now been replaced with
-DMECH=fullchem (default), -DMECH=Hg, or -DMECH=custom.

To use the new Hg mechanism with KPP, you must configure
GEOS-Chem with the -DMECH=Hg option.

Also note: right now the Hg simulation can only be used with
GEOS-Chem Classic.  A GCHP version may be available in the
near future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>