Move setup_wetscav call to avoid bug in first timestep
This bug fix removes the differences in H2O2 concentrations introduced in
wet scavenging when breaking up a GCHP run into multiple segments, and
fixes a problem limited to the first timestep of all runs. Before this fix
the call to setup_wetscav was before the first call to AIRQNT, when
State_Met%PMID was still zero. State_Met%C_H2O is calculated from PMID in
setup_wetscav, and thus was all zeros when wet scavenging was first called
in the run.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>