Update requirements for ReadTheDocs; use jinja2 prior to 3.1.0
A Python package update has broken backwards compatibility with the
ReadTheDocs theme versions prior to 1.0.0.  In order to fix this,
we now specify that we should use a version of jinja2 prior to 3.1.0.

Also, for safety's sake, request specific versions of other python
packages that have been proven to build the documentation properly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>