HEMCO submod update: Add clarifying RTD doc for EFYO time cycle option
This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/HEMCO repository:

0293219 (HEAD -> docs/dev, origin/docs/dev) Update RTD documentation in preparation for 3.5.2
2b8ad41 RTD update: Added clarifying text for EFYO time cycle flag option
03acc9b Merge branch 'main' into docs/dev
9a92575 Add info about EFYO switch to the HEMCO_Config.rc RTD chapter

This update adds some clarifying text to hopefully make it easier for
users to know when to change the time cycle flags from EFYO to CYS
(for restart files).  Version numbers in HEMCO have also been updated
to 3.5.2 and the CHANGELOG.md updated accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>