Make FindESMF.cmake use consistent libs
In `FindESMF.cmake` we spend a lot of time and effort defining the `ESMF_LIBRARIES` variable, including bringing in common libraries such as netCDF, MPI, and so on. However, we then hand-specify these later on for linking with ESMF. This patch makes two edits:

1. It includes the `dl` library in the ones which get filtered early on (basically just preventing duplication); and
2. It uses the `ESMF_LIBRARIES` variable at the end instead of re-specifying all of them again (which was causing libraries like `PIO` to be missed and, on at least one system, also `dl`).

This change should be 100% benign (zero-diff) and fixes issue #305.