Change default air pressure in advection to dry if not found in config file
GCHP.rc entry for USE_TOTAL_AIR_PRESSURE_IN_ADVECTION will override this
default.

Also in this commit: write log message if not correcting mass fluxes for
moisture. Previously a message was only written if correcting them. The
default if CORRECT_MASS_FLUX_FOR_HUMIDITY is not in GCHP.rc is to correct
them. This goes with the default of using dry air pressure in advection.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>