GEOS-Chem update: Merge PR #2593 (Prevent duration < frequency)
This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

4a10fac7e Merge PR #2593 (Prevent duration < frequency in History diags)

This PR fixes an issue where empty diagnostic files could be
created if the collection duration setting was less than
the frequency setting.  Now an error will be thrown and the
run will halt.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>