GEOS-Chem update: Merge PR #2689 (Block out rosenbrock code)
This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

189adce2c Merge PR #2689 (Block out rosenbrock_autoreduce code)

This PR adds code to automatically set a C-preprocessor switch
KPP_INTEGRATOR_AUTOREDUCE based on the name of the KPP integrator that
is being used. This is necessary so that we can block out code that is
specific to the rosenbrock_autoreduce integrator, which would result
in a compile-time error when we use other integrators.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>