GEOS-Chem update: Merge PR #2766 (KPP-Stadalone compilation fix)
This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

25db078e0 Merge PR #2766 (KPP-Standalone compilation fix)

This PR adds if blocks to the KPP/CMakeLists.txt file so that the
KPP-Standalone code will only be compiled if GEOS-Chem has been
configured with -DKPPSA=y.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>
GEOS-Chem update: Merge PR #2766 (KPP-Stadalone compilation fix)

This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

25db078e0 Merge PR #2766 (KPP-Standalone compilation fix)

This PR adds if blocks to the KPP/CMakeLists.txt file so that the
KPP-Standalone code will only be compiled if GEOS-Chem has been
configured with -DKPPSA=y.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>