GEOS-Chem update: Merge PR #2853 (Obspack I/O fix)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

e258f7850 Merge PR #2853 (Obspack now looks for "middle" in "time:comment")

PR #2853 adds code to ObsPack so that it will look for either
"middle of the averaging interval" or "midpoint of the averaging interval"
in the "time:comment" field.  This is used to determine whether the
timestamp of the observation should be placed at the start or middle
of the averaging interval.  This fixes the problem raised by @aschuh in
#2849 where ObsPack diagnostic output near the end of a UTC date were
zeroes.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>