GEOS-Chem update: Merge PR #2802 and PR #2869
This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

3817b551e Merge PR #2869 (Do not set DELP_DRY=0 if not in restart file)
32b543538 Merge PR #2802 (Add script to generate species tables for RTD doc)

The code for initializing DELP_DRY was moved from hco_interface_gc_mod.F90
to hco_utilities_gc_mod.F90 in PR #2521. This inadvertently led to
negative concentrations in simulations where DELP_DRY was not found in
the restart file. This is because DELP_DRY is updated in the first call
to AirQnt in hco_interfaces_gc_mod.F90. When it is not found in the
restart file, it is reset to zero which causes issues later in the
simulation (manifesting as negative values in mixing). To avoid these
issues, we remove the ELSE statement in PR #2869 setting DELP_DRY to
zero when it's not found in the restart file.


PR #2802 adds the script run/shared/rtd_species_per_simulation.py,
which generates ReST tables of all the species in a given GEOS-Chem
simulation. This can be used to easily update the ReadTheDocs
documentation when new species are added to GEOS-Chem.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>