HEMCO update: Merge PR #335 (Increase string length)
This commit informs the GCClassic superproject about the following
commits that were pushed to the GitHub geoschem/HEMCO repository:

3b81ba1 Merge PR #335 (Increase string length)

In PR #207 we increased the maximum string length to 5500 to allow
for global 0.25x0.3125 HEMCO standalone simulations. However for
0.125x0.15625 global grids, the length of the YEDGE and YMID string
in HEMCO_sa_Grid.rc are 12,103 and 10,657 characters, respectively.
To get around this we increased the length of string DUM from 5500
to 15000 in PR #335.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>