GEOS-Chem update: Merge PR #2959 (Bugfix for mass flux & cn scaling)
This commit informs the GCHP superproject about the following
commits that were pushed to the GitHub geoschem/geos-chem repository:

56f78f798 Merge PR #2969 (Bugfix/update gchp mass flux cn scalings)

PR #2959 builds on the mass flux and Courant number dynamic scaling
update submitted by Yuanjian Zhang (@yuanjianz) in #2959. It applies
a grid resolution scaling in addition to the timestep scaling previously
added. This makes the mass flux and Courant number scalings c
consistent with the custom version of MAPL currently used in the
model, specifically including the horizontal flux regridding fix
added in 14.4.2 in geoschem/MAPL#36.

This scaling will change once we update to MAPL 2.55, anticipated
for release in GCHP 14.7.0.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>